requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.1
djangorestframework==3.13.1
py-moneyed==2.0
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.3.0
